pyproject.toml
[project]
name = "wselfedu"
version = "0.6.0"
description = "Web server of the WSE project."
authors = [
    { name = "Sergei Mikurov", email = "454004@mail.ru" }
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "crispy-bootstrap5 (>=2025.6,<2026.0)",
    "dependency-injector (>=4.48.1,<5.0.0)",
    "dj-database-url (>=3.0.1,<4.0.0)",
    "django-bootstrap5 (>=25.1,<26.0)",
    "django-crispy-forms (>=2.4,<3.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "wse-exercises (<1.0.0)",
]


# Poetry
[tool.poetry]
package-mode = false

# Dependencies for static type checking
[tool.poetry.group.typecheck.dependencies]
django-stubs = "^5.2.1"
djangorestframework-stubs = "^3.16.0"
mypy = "^1.16.1"

# Dependencies for documenting project
[tool.poetry.group.docs.dependencies]
drf-spectacular = "^0.28.0"
furo = "^2024.8.6"
sphinx = "<8.2.0"
sphinx-copybutton = "^0.5.2"

# Dependencies for code style checking
[tool.poetry.group.lint.dependencies]
ruff = "^0.12.2"

# Dependencies for testing the project
[tool.poetry.group.test.dependencies]
pytest-cov = "^6.2.1"
pytest-django = "^4.11.1"
pytest-mock = "^3.14.1"


# Settings for building a Python package (ignore)
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"